# docs/swagger.yaml
definitions:
...
  response.AttachmentShort:
    properties:
      file_size:
        type: integer
      file_type:
        type: string
      file_url:
        type: string
      id:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.NoteResponse:
    properties:
      attachments:
        items:
          $ref: '#/definitions/response.AttachmentShort'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      related_ids:
        items:
          type: integer
        type: array
      summary:
        type: string
      tags:
        items:
          $ref: '#/definitions/response.TagShort'
        type: array
      title:
        type: string
      topic_id:
        type: integer
      updated_at:
        type: string
    type: object
  response.NotesListResponse:
    properties:
      notes:
        items:
          $ref: '#/definitions/response.NoteResponse'
        type: array
      total:
        type: integer
    type: object
...
  response.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  response.SummarizeResponse:
    properties:
      summary:
        type: string
    type: object
...
  response.TagShort:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  title: '---'
paths:
...
  /notes/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет заметку пользователя по id и все связанные вложения
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заметка успешно удалена
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Заметка не найдена NOTE_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка при удалении заметки DB_ERROR
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление заметки
      tags:
      - note
    get:
      consumes:
      - application/json
      description: Получение заметки пользователя по id
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заметка успешно получена
          schema:
            $ref: '#/definitions/response.NoteResponse'
        "404":
          description: Заметка не найдена NOTE_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получения заметки
      tags:
      - note
  /notes/{id}/archive:
    patch:
      consumes:
      - application/json
      description: Переносит заметку пользователя в архив (IsArchived = true)
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заметка архивирована
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Заметка не найдена NOTE_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка при архивировании заметки DB_ERROR
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Архивировать заметку
      tags:
      - note
  /notes/{id}/summarize:
    post:
      consumes:
      - application/json
      description: Генерирует краткое резюме для заметки пользователя по её ID
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Резюме успешно сгенерировано
          schema:
            $ref: '#/definitions/response.SummarizeResponse'
        "404":
          description: Заметка не найдена NOTE_NOT_FOUND
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка генерации резюме SUMMARY_ERROR, Ошибка сохранения резюме
            SUMMARY_SAVE_ERROR
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Суммаризация заметки по ID
      tags:
      - note
  /notes/create:
    post:
      consumes:
      - multipart/form-data
      description: Создаёт новую заметку пользователя с генерацией эмбеддинга, тегами
        и вложениями
      parameters:
      - description: Заголовок
        in: formData
        name: title
        required: true
        type: string
      - description: Содержимое
        in: formData
        name: content
        required: true
        type: string
      - collectionFormat: csv
        description: ID связанных заметок
        in: formData
        items:
          type: integer
        name: related_ids
        type: array
      - collectionFormat: csv
        description: ID тегов
        in: formData
        items:
          type: integer
        name: tag_ids
        type: array
      - collectionFormat: csv
        description: Вложения (image, audio, pdf)
        in: formData
        items:
          type: file
        name: attachments
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Заметка успешно создана
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать заметку
      tags:
      - note
  /notes/list:
    get:
      consumes:
      - application/json
      description: Выдаёт список всех заметок авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список заметок
          schema:
            $ref: '#/definitions/response.NotesListResponse'
        "500":
          description: 'Ошибка при получении заметок: DB_ERROR'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получения списка заметок
      tags:
      - note
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"